FIX: compiler error due use of long as input in main
Newer compiler versions complain about the 'long' type in the main
function for the argument counter

+++ taco/system/tools/devstate.c
@@ -69,7 +69,7 @@ void version(const char *cmd)
         exit(0);
 }
 
-int main(long argc,char **argv)
+int main(int argc,char **argv)
 {
 	int 		status,
 			i;
